Use getRectangles since layout isn't observable
Since the layout isn't observable, things like
`get blockLayouFeatures()` don't update if e.g. `layout.rectangles`
updates. With server-side rendering, this wasn't noticed since the
layout was already populated with rects when it got to the client side.
In preparation for rendering without the server-side component, this
change forces `get blockLayouFeatures()` to not cache its result the
first time it's evaluated, when it may have no rects.